Add animated sprite showcase to homepage
